[FIX] point_of_sale: balance pos session move when closing
Instances like deleting a tax used in a pos.order before closing a
session is allowed but it renders error when closing the session
because of imbalance move lines. In this commit, we are adding
a 'balancing' line in the move before it is created. This ensures
the closing of pos session. The introduced move line is named
'Difference at closing PoS session' and is a receivable making it
easy to find and can be reconciled.

closes odoo/odoo#62187

X-original-commit: e25d859cd1d24404891a5c73aeb3550217c8485a
Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>